requirements_test.txt
pytest==8.*
pytest-asyncio==0.*
pytest-homeassistant-custom-component==0.*
